requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
anyascii==0.3.1
    # via wagtail
asgiref==3.6.0
    # via django
astroid==2.15.0
    # via pylint
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via pytest
autopep8==2.0.2
    # via -r requirements.in
babel==2.12.1
    # via py-moneyed
beautifulsoup4==4.11.2
    # via wagtail
billiard==3.6.4.0
    # via celery
black==23.1.0
    # via -r requirements.in
boto3==1.26.89
    # via -r requirements.in
botocore==1.29.89
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2022.12.7
    # via
    #   django-money
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==7.2.1
    # via
    #   -r requirements.in
    #   django-coverage-plugin
    #   pytest-cov
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==39.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dill==0.3.6
    # via pylint
dj-stripe==2.7.3
    # via -r requirements.in
django==4.1.7
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-stripe
    #   django-allauth
    #   django-auto-prefetch
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-filter
    #   django-htmx
    #   django-modelcluster
    #   django-modeltranslation
    #   django-money
    #   django-permissionedforms
    #   django-rosetta
    #   django-storages
    #   django-taggit
    #   django-tex
    #   django-treebeard
    #   djangorestframework
    #   jsonfield
    #   wagtail
django-allauth==0.52.0
    # via -r requirements.in
django-analytical==3.1.0
    # via -r requirements.in
django-auto-prefetch==1.5.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-cleanup==7.0.0
    # via -r requirements.in
django-coverage-plugin==3.0.0
    # via -r requirements.in
django-crispy-forms==2.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-debug-toolbar==3.8.1
    # via -r requirements.in
django-filter==22.1
    # via wagtail
django-htmx==1.14.0
    # via -r requirements.in
django-modelcluster==6.0
    # via wagtail
django-modeltranslation==0.18.9
    # via -r requirements.in
django-money[exchange]==3.0.0
    # via -r requirements.in
django-permissionedforms==0.1
    # via wagtail
django-rosetta==0.9.8
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
django-taggit==3.1.0
    # via wagtail
django-tex==1.1.10
    # via -r requirements.in
django-treebeard==4.6.1
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.1
    # via -r requirements.in
faker==17.6.0
    # via factory-boy
flake8==6.0.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
hiredis==2.2.2
    # via -r requirements.in
html5lib==1.1
    # via wagtail
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via pylint
jinja2==3.1.2
    # via django-tex
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via dj-stripe
kombu==5.2.4
    # via celery
l18n==2021.3
    # via wagtail
lazy-object-proxy==1.9.0
    # via astroid
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy-extensions==1.0.0
    # via black
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.1
    # via wagtail
packaging==23.0
    # via
    #   black
    #   pytest
pathspec==0.11.0
    # via black
pdf2image==1.16.3
    # via -r requirements.in
pillow==9.4.0
    # via
    #   -r requirements.in
    #   pdf2image
    #   wagtail
platformdirs==3.1.1
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
polib==1.2.0
    # via django-rosetta
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
py-moneyed==2.0
    # via django-money
pyclean==2.7.0
    # via -r requirements.in
pycodestyle==2.10.0
    # via
    #   autopep8
    #   flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pyjwt[crypto]==2.6.0
    # via django-allauth
pylint==2.17.0
    # via
    #   -r requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.3
    # via -r requirements.in
pylint-plugin-utils==0.7
    # via pylint-django
pytest==7.2.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-reverse
    #   pytest-xdist
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-reverse==1.5.0
    # via -r requirements.in
pytest-xdist==3.2.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   celery
    #   django-modelcluster
    #   djangorestframework
    #   l18n
redis==4.5.1
    # via -r requirements.in
requests==2.28.2
    # via
    #   django-allauth
    #   django-rosetta
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.4
    # via beautifulsoup4
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stripe==4.2.0
    # via dj-stripe
tblib==1.7.0
    # via -r requirements.in
telepath==0.3
    # via wagtail
tomlkit==0.11.6
    # via pylint
typing-extensions==4.5.0
    # via
    #   django-modeltranslation
    #   py-moneyed
urllib3==1.26.14
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==4.2
    # via -r requirements.in
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
wrapt==1.15.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
